refactor: Aplica todas las correcciones del plan de refactorización 06-06-2026
- DOM-001 a DOM-007: Elimina Class1.cs stub, setters privados en
  TenantDomainEntity, cambia Enum.Parse por TenantStatus, DateTime.UtcNow,
  unifica PKs a int, crea ValueObjects (TenantCode, Email) y excepciones
  de dominio, migra TenantStatus enum a Application e Infrastructure.
- APP-001 a APP-012: Elimina stubs y código muerto (IPipelineBehavior
  custom, IValidator, IConfiguration, Interactor, Ports), implementa
  LoggingBehavior, unifica pipeline con MediatR, restaura DependencyInjection,
  hace public GeneralResponseDTO, init-only en CreateTenantCommand,
  usa IsInEnum en validator, enriquece ValidationException con metadatos,
  corrige ITenantRepository.GetByIdAsync con parámetro id.
- INF-001 a INF-011: Elimina stubs y tenant_setting.cs, implementa
  TenantRepository, propaga excepciones en ProvisioningService,
  asigna tenant_branding, sanitiza nombres de BD, usa TenantStatus enum,
  registra TenantDbContextFactory y MongoDbAudit en DI, completa
  Infrastructure/DependencyInjection.cs.
- IA-001 a IA-006: Restaura InterfaceAdapters/DependencyInjection.cs,
  elimina ICreateTenantInputPort del controller, elimina GeneralResponseDTO
  duplicado, renombra Presentes a Presenters, agrega ILogger en
  ExceptionHandlingMiddleware, valida null en TenantResolutionMiddleware.
- API-001 a API-005: Elimina AddValidatorsFromAssembly duplicado,
  aplica UseCors, agrega valores por defecto para Frontend config,
  consolida registros DI en Add{Layer}().

+++ src/ZooTech.Infrastructure/Tenant/TenantContext.cs
@@ -4,12 +4,12 @@ namespace ZooTech.Infrastructure.Tenant
 {
     public class TenantContext : ITenantContext
     {
-        public long TenantId {  get; private set; }
+        public int TenantId { get; private set; }
         public string Code { get; private set; } = "";
         public string SubDomain { get; private set; } = "";
         public string DatabaseName { get; private set; } = "";
 
-        public void SetTenant(long id, string code , string subDomain , string databaseName)
+        public void SetTenant(int id, string code, string subDomain, string databaseName)
         {
             TenantId = id;
             Code = code;
